refactor(padi): house port scan + sampler under src/ports/
No pre-existing port-forwarding subdir in padi (that feature is
@kolu/port-forward). Group detection next to ptyHost/ and terminalEndpoint/:
scan.ts, sampler.ts, tests, and a small barrel.
