+++ packages/padi/src/ports/scan.load.test.ts
@@ -1,5 +1,5 @@
 /**
- * Pin that `portScan.ts` loads under Node's real ESM resolver — the path
+ * Pin that `scan.ts` loads under Node's real ESM resolver — the path
  * production tsx/node takes. Vitest/esbuild rewrites named imports and will
  * happily leave a missing export as `undefined` until a call site runs; Node
  * throws `SyntaxError: does not provide export` at module evaluation. That
@@ -10,12 +10,12 @@ import { spawnSync } from "node:child_process";
 import { fileURLToPath } from "node:url";
 import { describe, expect, it } from "vitest";
 
-describe("portScan module loadability (Node ESM)", () => {
+describe("ports/scan module loadability (Node ESM)", () => {
   it("evaluates under node --import tsx without a missing-export SyntaxError", () => {
-    const portScanUrl = new URL("./portScan.ts", import.meta.url).href;
+    const portScanUrl = new URL("./scan.ts", import.meta.url).href;
     // Absolute path to this package's tsx (workspace-resolved), not PATH.
     const tsxLoader = fileURLToPath(
-      new URL("../node_modules/tsx/dist/loader.mjs", import.meta.url),
+      new URL("../../node_modules/tsx/dist/loader.mjs", import.meta.url),
     );
     const result = spawnSync(
       process.execPath,
@@ -29,7 +29,7 @@ describe("portScan module loadability (Node ESM)", () => {
       {
         encoding: "utf8",
         env: process.env,
-        cwd: fileURLToPath(new URL("..", import.meta.url)),
+        cwd: fileURLToPath(new URL("../..", import.meta.url)),
       },
     );
     expect(
