+++ packages/padi/src/terminalEndpoint/local.ts
@@ -49,7 +49,7 @@ import {
   createPortSampler,
   type PortSampler,
   type PortScanTarget,
-} from "../portSampler.ts";
+} from "../ports/sampler.ts";
 import { padiSurfaceCtx } from "../padiSurfaceCtx.ts";
 import { buildTerminalSpawnInput, ptyHostClient } from "../ptyHost/index.ts";
 import { notifyDirty } from "../publisher.ts";
